# public/swagger.yml
openapi: 3.0.3
info:
  title: RecTime API
  description: |
    RecTime のバックエンドAPIです。
    認証エンドポイントおよびイベントの書き込みAPIは、現時点では公開していません。
  version: 1.0.0

servers:
  - url: /api/v1
    description: APIのベースパス

paths:
  /students:
    get:
      tags: [Students]
      summary: 学生一覧を取得する
      responses:
        '200':
          description: 学生一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /students/{studentId}:
    get:
      tags: [Students]
      summary: 学生を取得する
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: 学生
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /events:
    get:
      tags: [Events]
      summary: イベント一覧を取得する
      parameters:
        - name: start_time
          in: query
          description: JSTの開始時刻（HHMM形式）で完全一致検索する
          schema:
            type: string
            pattern: '^\\d{4}$'
            example: '0930'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: イベント一覧
          content:
            application/json:
              schema:
                type: object
                required: [events, total, limit, offset]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

  /events/{eventId}:
    get:
      tags: [Events]
      summary: イベントを取得する
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: イベント
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /classes:
    get:
      tags: [Classes]
      summary: クラス一覧を取得する
      responses:
        '200':
          description: クラス一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassRoom'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gathering-spots:
    get:
      tags: [Gathering spots]
      summary: 集合場所一覧を取得する
      responses:
        '200':
          description: 集合場所一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GatheringSpot'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Gathering spots]
      summary: 集合場所を作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatheringSpotRequest'
      responses:
        '201':
          description: 作成した集合場所
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringSpot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gathering-groups:
    get:
      tags: [Gathering groups]
      summary: 対象グループ一覧を取得する
      responses:
        '200':
          description: 対象グループ一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GatheringGroup'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Gathering groups]
      summary: 対象グループを作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatheringGroupRequest'
      responses:
        '201':
          description: 作成した対象グループ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gathering-groups/{gatheringGroupId}/members:
    get:
      tags: [Gathering group members]
      summary: 対象グループの所属者一覧を取得する
      parameters:
        - $ref: '#/components/parameters/GatheringGroupId'
      responses:
        '200':
          description: 所属者一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GatheringGroupMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Gathering group members]
      summary: 対象グループへ所属者を追加する
      parameters:
        - $ref: '#/components/parameters/GatheringGroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGatheringGroupMemberRequest'
      responses:
        '201':
          description: 追加した所属情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringGroupMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: すでに対象グループへ所属している
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gathering-groups/{gatheringGroupId}/members/{userId}:
    delete:
      tags: [Gathering group members]
      summary: 対象グループから所属者を削除する
      parameters:
        - $ref: '#/components/parameters/GatheringGroupId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: 削除成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gatherings:
    get:
      tags: [Gatherings]
      summary: 集合予定一覧を取得する
      responses:
        '200':
          description: 集合予定一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gathering'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Gatherings]
      summary: 集合予定を作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatheringRequest'
      responses:
        '201':
          description: 作成した集合予定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gathering'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 対象グループにはすでに集合予定が登録されている
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /firebase-tokens:
    post:
      tags: [Firebase tokens]
      summary: Firebaseトークンを登録または更新する
      description: 学籍番号で学生を特定し、Firebaseトークンを登録または更新します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFirebaseTokenRequest'
      responses:
        '200':
          description: 登録または更新したFirebaseトークン
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterFirebaseTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notification-schedules:
    get:
      tags: [Notification schedules]
      summary: 通知予定一覧を取得する
      responses:
        '200':
          description: 通知予定一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationSchedule'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Notification schedules]
      summary: 通知予定を作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationScheduleRequest'
      responses:
        '201':
          description: 作成した通知予定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notifications/test:
    post:
      tags: [Notifications]
      summary: テスト用Firebase通知を送信する
      description: 環境変数 `TEST_FCM_TOKEN` に設定された端末へ通知を送信します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestNotificationRequest'
      responses:
        '200':
          description: 送信成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FcmNotificationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notifications/schedule/run:
    post:
      tags: [Notifications]
      summary: 期限を過ぎた通知予定を送信する
      description: 通常はcronで毎分実行されます。`now` はテスト・手動実行時のみ指定します。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                now:
                  type: string
                  format: date-time
                  example: '2026-07-16T09:00:00.000Z'
      responses:
        '200':
          description: 送信処理結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledNotificationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    GatheringGroupId:
      name: gatheringGroupId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 対象リソースが存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Student:
      type: object
      required:
        [student_id, user_id, user_name, class_room_id, attendance_number, student_id_number]
      properties:
        student_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        class_room_id:
          type: integer
        attendance_number:
          type: integer
        student_id_number:
          type: string

    Event:
      type: object
      required: [event_id, user_id, event_name, venue, start_time, end_time, created_at, updated_at]
      properties:
        event_id:
          type: integer
        user_id:
          type: integer
        event_name:
          type: string
        rule_text:
          type: string
          nullable: true
        venue:
          type: string
        start_time:
          type: string
          description: JSTの時刻（HHMM形式）
          example: '0930'
        end_time:
          type: string
          description: JSTの時刻（HHMM形式）
          example: '1745'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ClassRoom:
      type: object
      required: [f_class_room_id, f_class_code, f_name]
      properties:
        f_class_room_id:
          type: integer
        f_class_code:
          type: string
        f_name:
          type: string

    GatheringSpot:
      type: object
      required: [gathering_spot_id, gathering_spot_name, created_at, updated_at]
      properties:
        gathering_spot_id:
          type: integer
        gathering_spot_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateGatheringSpotRequest:
      type: object
      required: [gatheringSpotName]
      properties:
        gatheringSpotName:
          type: string
          minLength: 1

    GatheringGroup:
      type: object
      required: [gathering_group_id, gathering_group_name, created_at, updated_at]
      properties:
        gathering_group_id:
          type: integer
        gathering_group_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateGatheringGroupRequest:
      type: object
      required: [gatheringGroupName]
      properties:
        gatheringGroupName:
          type: string
          minLength: 1

    GatheringGroupMember:
      type: object
      required: [gathering_group_member_id, gathering_group_id, user_id, created_at, updated_at]
      properties:
        gathering_group_member_id:
          type: integer
        gathering_group_id:
          type: integer
        user_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    AddGatheringGroupMemberRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
          minimum: 1

    Gathering:
      type: object
      required:
        [gathering_id, gathering_group_id, event_id, gathering_spot_id, gathering_time, round, created_at, updated_at, gathering_group_name, event_name, gathering_spot_name]
      properties:
        gathering_id:
          type: integer
        gathering_group_id:
          type: integer
        event_id:
          type: integer
        gathering_spot_id:
          type: integer
        gathering_time:
          type: string
          description: HH:MM形式。`99:59`は集合時刻が未設定であることを表します。
          example: '08:45'
        round:
          type: integer
          minimum: 1
        gathering_group_name:
          type: string
        event_name:
          type: string
        gathering_spot_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateGatheringRequest:
      type: object
      required: [gatheringGroupId, eventId, gatheringSpotId]
      properties:
        gatheringGroupId:
          type: integer
          minimum: 1
        eventId:
          type: integer
          minimum: 1
        gatheringSpotId:
          type: integer
          minimum: 1
        gatheringTime:
          type: string
          pattern: '^(?:[01]\\d|2[0-3]):[0-5]\\d$|^99:59$'
          example: '08:45'
        round:
          type: integer
          minimum: 1
          maximum: 99

    RegisterFirebaseTokenRequest:
      type: object
      required: [studentNumber, platform]
      properties:
        studentNumber:
          type: string
          minLength: 1
        platform:
          type: integer
          enum: [1, 2]
          description: '`1`: iOS、`2`: Android'
        fcmToken:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
          deprecated: true
          description: '`fcmToken` の後方互換用の別名'
      anyOf:
        - required: [fcmToken]
        - required: [token]

    RegisterFirebaseTokenResponse:
      type: object
      required: [user, firebaseToken]
      properties:
        user:
          $ref: '#/components/schemas/NotificationTargetUser'
        firebaseToken:
          $ref: '#/components/schemas/FirebaseToken'

    NotificationTargetUser:
      type: object
      required: [user_id, user_name, is_live_active, created_at, updated_at]
      properties:
        user_id:
          type: integer
        user_name:
          type: string
        is_live_active:
          type: integer
          enum: [0, 1]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    FirebaseToken:
      type: object
      required:
        [firebase_token_id, user_id, platform, fcm_token, is_firebase_active, last_seen_at, created_at, updated_at]
      properties:
        firebase_token_id:
          type: integer
        user_id:
          type: integer
        platform:
          type: integer
          enum: [1, 2]
        fcm_token:
          type: string
        is_firebase_active:
          type: integer
          enum: [0, 1]
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    NotificationSchedule:
      type: object
      required:
        [notification_send_schedule_id, user_id, event_id, gathering_group_id, notification_id, importance, notification_type, title, body, send_status, send_at, created_at, updated_at]
      properties:
        notification_send_schedule_id:
          type: integer
        user_id:
          type: integer
          description: 通知予定を登録したユーザー
        event_id:
          type: integer
        gathering_group_id:
          type: integer
        notification_id:
          type: integer
        importance:
          type: integer
          minimum: 1
          maximum: 4
          default: 2
          description: '`1` が最も高く、`4` が最も低い重要度'
        notification_type:
          type: string
        title:
          type: string
        body:
          type: string
        send_status:
          type: string
          enum: [draft, sending, sent, failed]
        fcm_message_id:
          type: string
          nullable: true
        failed_reason:
          type: string
          nullable: true
        send_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateNotificationScheduleRequest:
      type: object
      required: [userId, eventId, gatheringGroupId, notificationId, sendAt]
      properties:
        userId:
          type: integer
          minimum: 1
        eventId:
          type: integer
          minimum: 1
        gatheringGroupId:
          type: integer
          minimum: 1
        notificationId:
          type: integer
          minimum: 1
        importance:
          type: integer
          minimum: 1
          maximum: 4
          default: 2
        sendAt:
          type: string
          format: date-time
          description: UTCオフセットを含むISO 8601形式
          example: '2026-07-16T09:00:00.000Z'

    TestNotificationRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1

    FcmNotificationResult:
      type: object
      required: [success, messageId]
      properties:
        success:
          type: boolean
          enum: [true]
        messageId:
          type: string

    ScheduledNotificationResult:
      type: object
      required: [checkedEvents, sent, failed]
      properties:
        checkedEvents:
          type: integer
          description: 今回確保した通知予定数
        sent:
          type: integer
          description: 送信に成功した端末数
        failed:
          type: integer
          description: 送信に失敗した通知予定数

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true

tags:
  - name: Students
    description: 学生情報
  - name: Events
    description: イベント
  - name: Classes
    description: クラス
  - name: Gathering spots
    description: 集合場所
  - name: Gathering groups
    description: 対象グループ
  - name: Gathering group members
    description: 対象グループの所属者
  - name: Gatherings
    description: 集合予定
  - name: Firebase tokens
    description: Firebaseトークン
  - name: Notification schedules
    description: 通知予定
  - name: Notifications
    description: Firebase通知